[CD-192] Rename: model to vae
